Enum panser::Error

pub enum Error {
    Bincode(Error),
    Cbor(Error),
    Eof,
    Generic(String),
    Io(Error),
    Json(Error),
    MsgpackDecode(Error),
    MsgpackEncode(Error),
    ParseInt(ParseIntError),
    Pickle(Error),
    TomlDecode(Error),
    TomlEncode(Error),
    Utf8(Utf8Error),
    UrlDecode(Error),
    UrlEncode(Error),
    Yaml(Error),
}

The error type for panser-releated operations and associated traits.

Errors mostly originate from the dependencies, but custom instances of Error can be crated with the Generic variant and a message.

Variants
Bincode(Error)

Decoding/encoding of the Bincode format failed.

Cbor(Error)

Decoding/encoding of the CBOR format failed.

Eof

End-of-File (EOF) reached.

Generic(String)

A generic or custom error occurred. The message should contain the detailed information.

Io(Error)

An I/O operation failed.

Json(Error)

Decoding/encoding of the JSON format failed.

MsgpackDecode(Error)

Decoding of the MessagePack format failed.

MsgpackEncode(Error)

Encoding of the MessagePack format failed.

ParseInt(ParseIntError)

Converting a string to an integer failed.

Pickle(Error)

Decoding/encoding of the Pickle format failed.

TomlDecode(Error)

Decoding of the TOML format failed.

TomlEncode(Error)

Encoding of the TOML format failed.

Utf8(Utf8Error)

A UTF8 operation failed.

UrlDecode(Error)

Decoding from a URL failed.

UrlEncode(Error)

Encoding from a URL failed.

Yaml(Error)

Decoding/encoding of the YAML format failed.

Methods

impl Error

pub fn code(&self) -> i32

Gets an error code related to the error.

This is useful as a return, or exit, code for a command line application, where a non-zero integer indicates a failure in the application. It can also be used for quickly and easily teseting equality between two errors.