Struct panser::Panser

pub struct Panser { /* fields omitted */ }

A Builder for transcoding.

Methods

impl Panser

pub fn new() -> Panser

Creates a new Panser with default options.

The defaults are stdin for input, stdout for output, JSON for the from format, Msgpack for the to format, no delimited input, no delimited output, no sized input, and no sized output. The Panser struct is implemented following the Builder pattern. Methods can be chained to change the defaults.

pub fn delimited_input(self, delimited: Option<&str>) -> Self

Sets a delimiter byte for the input and changes to framed reading of the data.

Data is read from the input source to the next delimiter byte. When the delimiter byte is reached, then all bytes up to the delimiter byte are transcoded. This continues until the End-of-File (EOF) is reached.

pub fn delimited_output(self, delimited: Option<&str>) -> Self

Sets a delimiter byte for the output.

The delimiter byte is appended to the output data.

pub fn from(self, from: Option<FromFormat>) -> Self

The format of the input.

If None, which is the default, then the input format is assumed to be JSON.

pub fn inputs(self, inputs: Option<Vec<&str>>) -> Self

The input source.

If None, which is the default, then stdin is used as the source. The value is a path to a file.

pub fn output(self, output: Option<&str>) -> Self

The output destination.

If None, which is the default, then stdout is used as the destination. The value is a path to a file.

pub fn radix(self, radix: Option<Radix>) -> Self

Sets the written output to be a space-separated list of bytes represented as numeric strings with a specific radix, or notation.

The data is still transcoded to the to format, but it is written to the output as a string. This is useful for debugging and creating an interactive console where humans are reading the serialized output.

pub fn run(self) -> Result<()>

Create a producer-consumer architecture for reading and writing data.

A separate thread is created and started for reading the input until End-of-File (EOF) is reached. If reading stdin, Ctrl+D can be used to force an EOF.

If input is None, then stdin is used for input. If output is None, then stdout is used for output. If the input (from) format is None, then the format is determined from the file extension if a file is provided and it has an extension. The default input format is JSON. If the input format is not JSON and a file with an appropriate extension is not used, then the from parameter should not be None. A similar procedure is used for the output (to) format.

pub fn sized_input(self, sized: bool) -> Self

Indicates the first four bytes is the total data length and changes to framed reading of the data.

The first four bytes are read as an unsigned 32-bit integer (u32) in Big Endian (Network Order). Then N number of bytes are read, where N is the size converted from the first four bytes. Once N bytes are read, all bytes up to the size are transcoded. This continues until the End-of-File (EOF) is reached.

pub fn sized_output(self, sized: bool) -> Self

Prepends the length of the data to the output.

The size of the output is prepended as an unsigned 32-bit integer (u32) in Big Endian (Network Order).

pub fn to(self, to: Option<ToFormat>) -> Self

The format of the output.

If None, which is the default, then the output format is assumed to be MessagePack (Msgpack).